Rename component names to camleCase format
