+++ src/app/components/articleCard.module.scss
