+++ src/app/components/newsList.module.scss
