+++ src/constant/index.ts
@@ -1,2 +1,2 @@
-export const API_KEY = 'a369778753444439853a99302b00788e';
+export const API_KEY = '63ab1c23401e406b8231dd55f56ed2e4';
 export const MAX_ARTICLE_PAGE = 20;
