+++ src/store-provider/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { newsApi } from "@/services/news-api";
+import { newsApi } from "@/services/newsApi";
 
 export const store = configureStore({
   reducer: {
