+++ src/types/categories.ts
@@ -1,4 +1,4 @@
-import CategoriesEnum from "./categories-enum";
+import CategoriesEnum from "./categoriesEnum";
 
 type Categories = CategoriesEnum.Business | CategoriesEnum.Entertainment | CategoriesEnum.General | CategoriesEnum.Health | CategoriesEnum.Science | CategoriesEnum.Sports | CategoriesEnum.Technology | '';
