+++ src/types/categoriesEnum.ts
