+++ src/types/index.ts
@@ -1,7 +1,7 @@
 export type { default as Article } from './article';
-export type { default as NewsApiResponse } from './news-api-response';
-export type { default as SearchInType } from './search-in-type';
-export type { default as NewsListFilters } from './news-list-filters';
+export type { default as NewsApiResponse } from './newsApiResponse';
+export type { default as SearchInType } from './searchInType';
+export type { default as NewsListFilters } from './newsListFilters';
 export type { default as Categories } from './categories';
-export { default as CategoriesEnum } from './categories-enum';
-export { default as SearchInEnum } from './search-in-enum';
+export { default as CategoriesEnum } from './categoriesEnum';
+export { default as SearchInEnum } from './searchInEnum';
