+++ src/types/newsApiResponse.ts
