+++ src/types/searchInEnum.ts
